backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
django-import-export==3.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.1.0
pandas==2.1.3
openpyxl==3.1.2
python-docx==1.1.0
reportlab==4.0.7
weasyprint==60.1
pydyf==0.10.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
numpy==1.25.2
matplotlib==3.8.1